Avoid importing Network type from SDK
