+++ src/components/contextual/pages/vebal/LMVoting/GaugesFilters.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { networkLabelMap } from '@/composables/useNetwork';
-import { Network } from '@balancer-labs/sdk';
-import { ref, computed } from 'vue';
+import { Network } from '@/lib/config';
 
 /**
  * TYPES
