+++ src/composables/useNetwork.ts
@@ -1,22 +1,7 @@
-import config from '@/lib/config';
+import config, { Network } from '@/lib/config';
 import { configService } from '@/services/config/config.service';
 import { RouteParamsRaw } from 'vue-router';
 import { Config } from '@/lib/config/types';
-
-// We don't import Network from sdk to avoid extra bundle size when loading app (while the SDK is not tree-shakable)
-export enum Network {
-  MAINNET = 1,
-  ROPSTEN = 3,
-  RINKEBY = 4,
-  GOERLI = 5,
-  GÖRLI = 5,
-  OPTIMISM = 10,
-  KOVAN = 42,
-  GNOSIS = 100,
-  POLYGON = 137,
-  ARBITRUM = 42161,
-  FANTOM = 250,
-}
 
 /**
  * STATE
