+++ src/pages/claim/index.vue
@@ -25,7 +25,7 @@ import { BalanceMap } from '@/services/token/concerns/balances.concern';
 import useWeb3 from '@/services/web3/useWeb3';
 import { TOKENS } from '@/constants/tokens';
 import { buildNetworkIconURL } from '@/lib/utils/urls';
-import { Network } from '@balancer-labs/sdk';
+import { Network } from '@/lib/config';
 
 /**
  * TYPES
