+++ src/services/cowswap/cowswapProtocol.service.ts
@@ -1,4 +1,4 @@
-import { Network } from '@balancer-labs/sdk';
+import { Network } from '@/lib/config';
 import axios from 'axios';
 
 import { networkId } from '@/composables/useNetwork';
