+++ src/services/web3/connectors/trustwallet/walletconnect.connector.ts
@@ -3,7 +3,7 @@ import WalletConnectProvider from '@walletconnect/web3-provider';
 import ConfigService from '@/services/config/config.service';
 import config from '@/lib/config';
 import { WalletError } from '@/types';
-import { Network } from '@balancer-labs/sdk';
+import { Network } from '@/lib/config';
 import { Connector, ConnectorId } from '../connector';
 import { Config } from '@/lib/config/types';
